move adapters to own package
